Implement a config system and a CLI
The config system is an extension of the kamaki config system,
tailored to support configuration files with multiple clouds and
multiple syncs. A sync contains all the attributes needed to sync
a local directory with a remote container, included a reference to
a cloud.

The CLI is interactive and can be used to start, pause and control
the syncing process.