Emancipate GUI when retrieving endpoints
This patch is related to the endpoint URLs needed to make various
UI operations i.e., login and vising the pithos page on a browser.

The helper used to provide the GUI with all URLs necessary, while
now they are retrieved by the GUI itself, with the tokenless
request "POST <auth_url endpoint>/tokens".