Create and save settings is they are missing
If some settings are missing (e.g., not found in the configuration
file), agkyra GUI will run and let the user complete them through
the "setting" menu option. If the default_sync, sync or cloud
options are missing, they are created automatically.

Settings are saved only if the user opens the "settings" window.