Modify helper class to support multiple clients
The class is renamed to "SessionHelper" and moved to
"agkyra.protocol". It uses the session database to get information
about the status of the protocol server.

The session database is used to store the address and ui_id of the
protocol server as well as a time stamp. The helper uses this
information to resolve if the protocol server is active. If the
server is active, the helper assumes there is a session running and
allows the client to use the session credentials to connect to it.
Otherwise, it generates new session credentials and offers the
ability to start a protocol server.

The GUI class is also modified to raise an assertion error if
another GUI is already running. This is resolved by comparing the
session information stored in the session file with the one
provided by the helper. If they match, a GUI is running, otherwise
a new GUI is launched.