On SSL error, fall back to default certificates
Create a decorator "ssl_fall_back" which catches a KamakiSSLError,
patches kamaki with the default certificates bundle and reruns the
failed method.

The decorator is used on the "SyncerSettings._get_pithos_client"
method, because this is the first point of SSL failure.

The default certificates bundle is acquired from the package
"certifi", which has been added to dependencies.