Launch the server in a different process
To achieve this, a new method "wait_session_to_load" is
implemented in the SessioHelper class, which blocks the execution
until a server session (any server session) is loaded.

The SessionHelper does not create a new session on instantiation.
Now the caller has to explicitly ask the helper to load a session
or create a new one.