Move heartbeat update operations to main session
Instead of updating the "heart" database from the context of each
connection, update it from the context of the helper session.
The connections will still poll the database to decide whether
they are going to stay open and still have the authority to empty
the database when there is a shutdown request.