Use common status codes across UI components
All status codes are defined in an external JSON file named
"ui_common.json". This file is loaded by the UI components if they
must exchange statuses.

Also, the protocol for propagating the status of the server has
been modified. Instead of using separate flags to denote if the
server is in some status or not, we now use a status code which
must be defined in the common JSON file.

In the present commit, the WebSocket server is managing status
codes in that fashion, and the GUI is able to understand the
responses and adjust its behavior accordingly.