Help GUI user to resolve folder/container problems
Implement a dialogue window which is pop up when the directory or
container have been removed. The proposed solution is to create
the missing piece and reset the syncing process. If the user
accepts, the proposed solution is enforced. Otherwise the system
stays in a non working mode, where the user can fix the problem by
handling the settings.