Manage GUI launching in cli.py, handle CLI ctrl-C
When starting GUI from CLI, the user may stop it with a keyboard
interrupt. We treat this as equivalent to quitting the application
through the GUI, therefore the interrupt is caught and handled
accordingly.

The implementation required to move some launching code from
"gui.py" to "cli.py", which enabled the reuse of some CLI
methods (_start and _shutdown).